+++ test/clojure_lsp/api_test.clj
@@ -84,15 +84,15 @@
       (let [result (api/diagnostics {:project-root (io/file "integration-test/sample-test")
                                      :namespace '[sample-test.api.diagnostics.a]
                                      :raw? true})]
-        (is (= 1 (:result-code result)))
+        (is (= 3 (:result-code result)))
         (is (= "src/sample_test/api/diagnostics/a.clj:2:0: error: [unresolved-symbol] Unresolved symbol: some-unknown-var" (:message result)))
         (is (= 1 (count (:diagnostics result))))))
-    (testing "unused-public-var custom lint fn"
+    (testing "unused-public-var custom lint fn returning only info"
       (reset! db/db {})
       (let [result (api/diagnostics {:project-root (io/file "integration-test/sample-test")
                                      :namespace '[sample-test.api.diagnostics.d]
                                      :raw? true})]
-        (is (= 1 (:result-code result)))
+        (is (= 0 (:result-code result)))
         (is (= "src/sample_test/api/diagnostics/d.clj:2:6: info: [unused-public-var] Unused public var 'sample-test.api.diagnostics.d/unused-public-var'" (:message result)))
         (is (= 1 (count (:diagnostics result))))))
     (testing "when namespace does not exists"
@@ -109,7 +109,7 @@
                                      :output {:canonical-paths true}
                                      :namespace '[sample-test.api.diagnostics.a]
                                      :raw? true})]
-        (is (= 1 (:result-code result)))
+        (is (= 3 (:result-code result)))
         (is (= (format "%s:2:0: error: [unresolved-symbol] Unresolved symbol: some-unknown-var"
                        (.getCanonicalPath (io/file "integration-test/sample-test/src/sample_test/api/diagnostics/a.clj")))
                (:message result)))