+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   - Support renaming namespaces as well with `rename` feature.
   - Use relative paths instead of absolute paths on diff messages.
   - Add `analyze-project!` to analyze whole project only and cache analysis, useful for REPL usage for example. 
+  - Follow same exit status from clj-kondo for `diagnostics` feature. #572
 
 ## 2021.09.13-22.25.35